requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_tflow_select=2.1.0=gpu
absl-py=0.7.1=py36_0
astor=0.7.1=py36_0
blas=1.0=mkl
c-ares=1.15.0=h7b6447c_1
ca-certificates=2019.5.15=0
certifi=2019.6.16=py36_1
cloudpickle=1.2.1=pypi_0
cudatoolkit=10.1.168=0
cudnn=7.6.0=cuda10.1_0
cupti=10.1.168=0
cycler=0.10.0=py36_0
dbus=1.13.6=h746ee38_0
expat=2.2.6=he6710b0_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
future=0.17.1=pypi_0
gast=0.2.2=py36_0
glib=2.56.2=hd408876_0
google-pasta=0.1.7=py_0
grpcio=1.16.1=py36hf8bcb03_1
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
gym=0.14.0=pypi_0
h5py=2.9.0=py36h7918eee_0
hdf5=1.10.4=hb1b8bf9_0
icu=58.2=h9c2bf20_1
intel-openmp=2019.4=243
jpeg=9b=h024ee3a_2
keras-applications=1.0.8=py_0
keras-preprocessing=1.1.0=py_1
kiwisolver=1.1.0=py36he6710b0_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.8.0=hd408876_0
libstdcxx-ng=9.1.0=hdf63c60_0
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
markdown=3.1.1=py36_0
matplotlib=3.1.0=py36h5429711_0
mkl=2019.4=243
mkl-service=2.0.2=py36h7b6447c_0
mkl_fft=1.0.12=py36ha843d7b_0
mkl_random=1.0.2=py36hd81dba3_0
ncurses=6.1=he6710b0_1
numpy=1.16.4=py36h7e9f1db_0
numpy-base=1.16.4=py36hde5b4d6_0
openssl=1.1.1c=h7b6447c_1
pcre=8.43=he6710b0_0
pip=19.1.1=py36_0
protobuf=3.8.0=py36he6710b0_0
pyglet=1.3.2=pypi_0
pymdptoolbox=4.0b3=pypi_0
pyparsing=2.4.0=py_0
pyqt=5.9.2=py36h05f1152_2
python=3.6.9=h265db76_0
python-dateutil=2.8.0=py36_0
pytz=2019.1=py_0
qt=5.9.7=h5867ecd_1
readline=7.0=h7b6447c_5
scipy=1.3.0=py36h7c811a0_0
setuptools=41.0.1=py36_0
sip=4.19.8=py36hf484d3e_0
six=1.12.0=py36_0
sqlite=3.29.0=h7b6447c_0
tensorboard=1.14.0=py36hf484d3e_0
tensorflow=1.14.0=gpu_py36h3fb9ad6_0
tensorflow-base=1.14.0=gpu_py36he45bfe2_0
tensorflow-estimator=1.14.0=py_0
tensorflow-gpu=1.14.0=h0d30ee6_0
termcolor=1.1.0=py36_1
tk=8.6.8=hbc83047_0
tornado=6.0.3=py36h7b6447c_0
werkzeug=0.15.4=py_0
wheel=0.33.4=py36_0
wrapt=1.11.2=py36h7b6447c_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
